fix coadd3d scripts
